Separate text_processors  to avoid circular ref
